Clean architecture workspaces refactor
