+++ src/backend/libs/modules/FlowJudge.Workspaces.Application.Abstractions/Commands/UpdateWorkspaceCommand.cs
@@ -1,6 +1,6 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 
-namespace FlowJudge.Workspaces.Application.Commands
+namespace FlowJudge.Workspaces.Application.Abstractions.Commands
 {
     public sealed record UpdateWorkspaceCommand : ICommand
     {
