+++ src/backend/libs/modules/FlowJudge.Workspaces.Application.Abstractions/ErrorCodes.cs
@@ -1,0 +1,9 @@
+﻿namespace FlowJudge.Workspaces.Application.Abstractions
+{
+    public static class ErrorCodes
+    {
+        public const string WorkspaceCreateFailed = "workspace.create_failed";
+        public const string WorkspaceUpdateFailed = "workspace.update_failed";
+        public const string WorkspaceNotFound = "workspace.not_found";
+    }
+}
