+++ src/backend/libs/modules/FlowJudge.Workspaces.Application.Abstractions/FlowJudge.Workspaces.Application.Abstractions.csproj
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\FlowJudge.Common.Sql\FlowJudge.Common.Sql.csproj" />
-    <ProjectReference Include="..\FlowJudge.Workspaces.Application.Abstractions\FlowJudge.Workspaces.Application.Abstractions.csproj" />
+    <ProjectReference Include="..\..\FlowJudge.Common.Application\FlowJudge.Common.Application.csproj" />
     <ProjectReference Include="..\FlowJudge.Workspaces.Domain\FlowJudge.Workspaces.Domain.csproj" />
   </ItemGroup>
