+++ src/backend/libs/modules/FlowJudge.Workspaces.Application.Abstractions/Queries/GetUserWorkspacesQuery.cs
@@ -1,8 +1,8 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 using FlowJudge.Common.Utils.Pagination;
-using FlowJudge.Workspaces.Domain.Workspace.ReadModels;
+using FlowJudge.Workspaces.Application.Abstractions.Models;
 
-namespace FlowJudge.Workspaces.Application.Queries
+namespace FlowJudge.Workspaces.Application.Abstractions.Queries
 {
     public sealed record GetUserWorkspacesQuery(Guid UserId, PageQuery Pagination)
         : IQuery<PagedList<WorkspaceListItem>>;
