+++ src/backend/libs/modules/FlowJudge.Workspaces.Application/Commands/CreateWorkspaceCommandHandler.cs
@@ -3,8 +3,10 @@ using FlowJudge.Common.Application.Transactional;
 using FlowJudge.Common.Domain;
 using FlowJudge.Common.Sql.UnitOfWork;
 using FlowJudge.Common.Utils.Time;
+using FlowJudge.Workspaces.Application.Abstractions;
+using FlowJudge.Workspaces.Application.Abstractions.Commands;
+using FlowJudge.Workspaces.Application.Abstractions.Ports;
 using FlowJudge.Workspaces.Domain.Workspace.Model;
-using FlowJudge.Workspaces.Infrastructure;
 
 namespace FlowJudge.Workspaces.Application.Commands
 {
@@ -40,7 +42,7 @@ namespace FlowJudge.Workspaces.Application.Commands
             }
             catch (Exception ex)
             {
-                return ApplicationResultFactory.Failure<Guid>(ex.Message, "workspace.create_exception");
+                return ApplicationResultFactory.Failure<Guid>(ex.Message, ErrorCodes.WorkspaceCreateFailed);
             }
         }
     }
