--- src/backend/libs/modules/FlowJudge.Workspaces.Application/ErrorCodes.cs
@@ -1,7 +1,0 @@
-﻿namespace FlowJudge.Workspaces.Application
-{
-    internal static class ErrorCodes
-    {
-        public const string WorkspaceNotFound = "workspace.not_found";
-    }
-}
