+++ src/backend/libs/modules/FlowJudge.Workspaces.Application/FlowJudge.Workspaces.Application.csproj
@@ -9,7 +9,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\FlowJudge.Common.Application\FlowJudge.Common.Application.csproj" />
     <ProjectReference Include="..\..\FlowJudge.Common.Domain\FlowJudge.Common.Domain.csproj" />
-    <ProjectReference Include="..\FlowJudge.Workspaces.Infrastructure\FlowJudge.Workspaces.Infrastructure.csproj" />
+    <ProjectReference Include="..\FlowJudge.Workspaces.Application.Abstractions\FlowJudge.Workspaces.Application.Abstractions.csproj" />
   </ItemGroup>
 
 </Project>
