+++ src/backend/libs/modules/FlowJudge.Workspaces.Infrastructure/Repositories/Workspaces/Mappers/WorkspaceMapper.cs
@@ -1,5 +1,5 @@
-﻿using FlowJudge.Workspaces.Domain.Workspace.Model;
-using FlowJudge.Workspaces.Domain.Workspace.ReadModels;
+﻿using FlowJudge.Workspaces.Application.Abstractions.Models;
+using FlowJudge.Workspaces.Domain.Workspace.Model;
 using FlowJudge.Workspaces.Infrastructure.Repositories.Workspaces.DbModel;
 
 namespace FlowJudge.Workspaces.Infrastructure.Repositories.Workspaces.Mappers
@@ -56,7 +56,7 @@ namespace FlowJudge.Workspaces.Infrastructure.Repositories.Workspaces.Mappers
             return (workspace, members);
         }
 
-        public static WorkspaceListItem ToDomainModel(this WorkspaceListItemDbModel dbModel)
+        public static WorkspaceListItem ToModel(this WorkspaceListItemDbModel dbModel)
         {
             return new WorkspaceListItem
             {
